new "configure" script; use yarn workspaces. Refs STCOR-140
`util/configure/` builds a whole environment in one go after cloning
stripes-core.

Added a section about configure the `@folio` npm registry to use the CI
registry because (A) that's where `stripes-cli` is currently published
and there's no reason to build it from source (as was described in the
previous version) unless you're actively developing it and (B) it'll
bring your environment closer into line with CI, which should help make
tests more reproducible. Consistency is a good thing.