STRIPES-723 add RXJS upgrade to Juniper roadmap (#116)
Increment `rxjs` from `v5` to `v6`. This will require changes at least
in `stripes-core` and `stripes-connect` and probably other apps as well
(`ui-eholdings`).

If the version needs to be synced across all modules, the dependency
should be moved up to the platform level. If it only needs to be synced
across the stripes libraries, i.e. if apps' use is internal, I _think_
we would be fine supplying it to stripes-* via `@folio/stripes` but
we'll need to test that.

In either case, if we are changing a peer dependency, this work will
need to be part of a major release.

Refs STRIPES-723