Add quotes around "@folio:registry"
Without quotes around "@folio:registry", this operation fails on Windows (PowerShell).  Instead of creating the expected config line, it creates a config with the URL as the key, assigned to 'true'.

PS> yarn config list
yarn config v1.22.19
info yarn config
{
  ...
  'https://repository.folio.org/repository/npm-folio/': true,
  ...
}

With the quotes, the config statement works as expected.